cyber_risk_platform/requirements.txt
flask
python-nmap
requests
fpdf
streamlit
pandas
plotly
